Fix size_t/uint64_t types (#586)
* Remove uint32 in favor of uint32_t for better compatibility

* Switch size_t to uint32_t and uint64_t

* Clang-format

* Remove bad edits

+++ src/core/assets/AssetManager.cc
@@ -107,11 +107,11 @@ namespace sp {
 
     std::vector<std::string> AssetManager::ListBundledAssets(std::string_view prefix,
         std::string_view extension,
-        size_t maxDepth) const {
+        uint32_t maxDepth) const {
         std::vector<std::string> results;
         std::optional<std::filesystem::path> ignorePrefix;
-        std::function<void(std::filesystem::path, const std::filesystem::path &, size_t)> appendDirectoryFiles =
-            [&](auto dir, auto &base, size_t depth) {
+        std::function<void(std::filesystem::path, const std::filesystem::path &, uint32_t)> appendDirectoryFiles =
+            [&](auto dir, auto &base, uint32_t depth) {
                 if (std::filesystem::is_directory(dir)) {
                     for (auto &entry : std::filesystem::directory_iterator(dir)) {
                         if (entry.is_regular_file() && (extension.empty() || entry.path().extension() == extension)) {
@@ -133,7 +133,7 @@ namespace sp {
         if (!bundleIndex.empty()) {
             for (auto &[path, range] : bundleIndex) {
                 if (starts_with(path, prefix)) {
-                    size_t depth = 0;
+                    uint32_t depth = 0;
                     size_t sep = prefix.length();
                     while (depth <= maxDepth && sep != std::string::npos) {
                         sep = path.find('/', sep + 1);
